# specs/007-crud-departamentos-roles/contracts/departamentos-permisos.openapi.yaml
openapi: 3.0.3
info:
  title: API Departamentos (Referencia para UI con permisos por rol)
  version: 1.0.0
  description: >-
    Contrato de referencia para el módulo visual de departamentos con autenticación
    HTTP Basic y diferenciación de permisos por rol en frontend.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/departamentos:
    get:
      summary: Listar departamentos
      security:
        - basicAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Lista paginada de departamentos (10 por consulta)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDepartamentosResponse"
        "401":
          description: Sesión no válida o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      summary: Crear departamento
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartamentoCreateRequest"
      responses:
        "201":
          description: Departamento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartamentoResponse"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Sesión no válida o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Operación no autorizada para el rol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflicto de negocio (por ejemplo nombre duplicado)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /api/v1/departamentos/{id}:
    get:
      summary: Obtener departamento por id
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Departamento encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartamentoResponse"
        "401":
          description: Sesión no válida o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Departamento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      summary: Actualizar departamento
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartamentoUpdateRequest"
      responses:
        "200":
          description: Departamento actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartamentoResponse"
        "400":
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Sesión no válida o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Operación no autorizada para el rol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Departamento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflicto de negocio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      summary: Eliminar departamento
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Departamento eliminado
        "401":
          description: Sesión no válida o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Operación no autorizada para el rol
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Departamento no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Departamento no eliminable por reglas de negocio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    DepartamentoCreateRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100

    DepartamentoUpdateRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100

    DepartamentoResponse:
      type: object
      required:
        - id
        - nombre
        - empleadosCount
      properties:
        id:
          type: integer
          minimum: 1
        nombre:
          type: string
          maxLength: 100
        empleadosCount:
          type: integer
          minimum: 0

    PaginatedDepartamentosResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/DepartamentoResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          enum: [10]
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0

    ApiError:
      type: object
      required:
        - message
      properties:
        message:
          type: string